Make vendor delete more OOP
