+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -1,89 +1,18 @@
 package seedu.address.logic.commands;
 
-import static java.util.Objects.requireNonNull;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_VENDOR;
-
-import java.util.List;
-
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.ToStringBuilder;
-import seedu.address.logic.Messages;
-import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.Model;
-import seedu.address.model.vendor.Vendor;
 
 /**
  * Deletes a vendor identified using it's displayed index from the address book.
  */
-public class DeleteCommand extends Command {
+public abstract class DeleteCommand extends Command {
 
     public static final String COMMAND_WORD = "delete";
+    protected final Index targetIndex;
 
-    /**
-     * Enum to specify the type of item to delete.
-     */
-    public enum ItemTypeToDelete {
-        VENDOR,
-    }
-
-    public static final String MESSAGE_USAGE = COMMAND_WORD
-            + ": Deletes the person identified by the index number used in the displayed vendor list.\n"
-            + "Parameters: "
-            + PREFIX_VENDOR
-            + "INDEX (must be a positive integer)\n"
-            + "Example: " + COMMAND_WORD + " " + PREFIX_VENDOR + "1";
-
-    public static final String MESSAGE_DELETE_VENDOR_SUCCESS = "Deleted Vendor: %1$s";
-
-    // The index of item to be deleted
-    private final Index targetIndex;
-    private final ItemTypeToDelete itemType;
-
-    /**
-     * Creates a DeleteCommand to delete the specified item.
-     */
-    public DeleteCommand(ItemTypeToDelete itemType, Index targetIndex) {
+    public DeleteCommand(Index targetIndex) {
         this.targetIndex = targetIndex;
-        this.itemType = itemType;
-    }
-
-    @Override
-    public CommandResult execute(Model model) throws CommandException {
-        requireNonNull(model);
-
-        if (itemType == ItemTypeToDelete.VENDOR) {
-            return deleteVendor(model);
-        } else {
-            throw new CommandException(Messages.MESSAGE_INVALID_COMMAND_FORMAT);
-        }
-    }
-
-    private CommandResult deleteVendor(Model model) throws CommandException {
-        List<Vendor> lastShownList = model.getFilteredVendorList();
-
-        if (targetIndex.getZeroBased() >= lastShownList.size()) {
-            throw new CommandException(Messages.MESSAGE_INVALID_VENDOR_DISPLAYED_INDEX);
-        }
-
-        Vendor vendorToDelete = lastShownList.get(targetIndex.getZeroBased());
-        model.deleteVendor(vendorToDelete);
-        return new CommandResult(String.format(MESSAGE_DELETE_VENDOR_SUCCESS, Messages.format(vendorToDelete)));
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        if (other == this) {
-            return true;
-        }
-
-        // instanceof handles nulls
-        if (!(other instanceof DeleteCommand)) {
-            return false;
-        }
-
-        DeleteCommand otherDeleteCommand = (DeleteCommand) other;
-        return targetIndex.equals(otherDeleteCommand.targetIndex)
-                && itemType.equals(otherDeleteCommand.itemType);
     }
 
     @Override
